go/runtime/bundle: Move component to a new file
